[UPDATE] fix enroll test

+++ backend/internals/services/enroll.go
@@ -1,5 +1,5 @@
 package services
 
 type EnrollServices interface {
-	EnrollUser(userId uint64, courseId uint64) error
+	EnrollUser(userId uint, courseId uint64) error
 }
