+++ backend/internals/services/enroll_service.go
@@ -15,7 +15,7 @@ func NewEnrollService(enrollRepo repositories.EnrollRepo) *EnrollService { // Ac
 }
 
 // EnrollUser enrolls a user in a course
-func (s *EnrollService) EnrollUser(userId uint64, courseId uint64) error {
+func (s *EnrollService) EnrollUser(userId uint, courseId uint64) error {
 	// Use the EnrollUser method from the repository to enroll the user
 	err := s.enrollRepo.EnrollUser(userId, courseId)
 	if err != nil {
